Move some significant components
Now components that are related to pages placed in
/components/pageRelated folder.
